Moved an example class to an independent file
